+++ pages/modals/aud_abmauditores.php
@@ -1,0 +1,46 @@
+<div class="modal fade" id="modal-abm-auditor">
+    <div class="modal-dialog">
+    <div class="modal-content">
+        <div class="modal-header">
+            <button type="button" class="close" data-dismiss="modal"
+                aria-label="Close">
+                <span aria-hidden="true">&times;</span></button>
+            <h2 class="modal-title" id='modal-abm-auditor-title'>Nuevo Auditor</h2>
+        </div>
+        <div class="modal-body">
+            <!-- form start -->
+            <form method="post" role="form" action="">
+                
+                <input type="hidden" class="form-control" name="idente" placeholder="id_ente" id='modal-abm-auditor-id-ente' >
+                <input type="hidden" class="form-control" name="id" placeholder="id" id='modal-abm-auditor-id' >
+                
+                <div class="box-body">
+                    <div class="form-group">
+                        <label for="nombre">Nombre</label>
+                        <input type="text" class="form-control" name="nombre" placeholder="Nombre" id='modal-abm-auditor-nombre' required>
+                    </div>
+                    <div class="form-group">
+                        <label for="apellido">Apellido</label>
+                        <input type="text" class="form-control" name="apellido" placeholder="Apellido" id='modal-abm-auditor-apellido' required>
+                    </div>
+                    <div class="form-group">
+                        <label for="dni">DNI</label>
+                        <input type="text" class="form-control" name="dni" placeholder="DNI" id='modal-abm-auditor-dni'>
+                    </div>                    
+                    <div class="form-group">
+                        <div class="col-sm-8"></div>
+                        <div class="col-sm-2">
+                            <input type="button" name="AddAuditor" class="btn  btn-raised btn-success" value="Guardar" id='modal-abm-auditor-submit'>
+                        </div>
+                        <div class="col-sm-2">
+                            <button type="button" class="btn btn-default pull-left" data-dismiss="modal">Cancelar</button>
+                        </div>
+                    </div>
+                </div>
+            </form>
+        </div>
+    </div>
+        <!-- /.modal-content -->
+    </div>
+    <!-- /.modal-dialog -->
+</div>